mrege perp view refactor
