+++ callisto_core/delivery/wizard.py
@@ -1,56 +1,22 @@
 import logging
 
-from wizard_builder.view_helpers import StepsHelper, StorageHelper
-from wizard_builder.views import WizardView
+from wizard_builder import views as wizard_builder_views
 
-from django.core.urlresolvers import reverse
 from django.http import HttpResponse
 
-from .views import ReportActionView, ReportUpdateView, SecretKeyStorageHelper
+from . import view_helpers, view_partials
 
 logger = logging.getLogger(__name__)
 
 
-class ReportStepsHelper(StepsHelper):
-
-    def url(self, step):
-        return reverse(
-            self.view.request.resolver_match.view_name,
-            kwargs={
-                'step': step,
-                'uuid': self.view.report.uuid,
-            },
-        )
-
-
-class EncryptedStorageHelper(
-    SecretKeyStorageHelper,
-    StorageHelper,
-):
-
-    @property
-    def report(self):
-        return self.view.report
-
-    def current_data_from_storage(self):
-        if self.secret_key:
-            return self.report.decrypted_report(self.secret_key).get('data', {})
-        else:
-            return {'data': {}}
-
-    def add_data_to_storage(self, data):
-        data = {'data': data}
-        self.report.encrypt_report(data, self.secret_key)
-
-
 class EncryptedWizardView(
-    ReportUpdateView,
-    WizardView,
+    view_partials.ReportUpdateView,
+    wizard_builder_views.WizardView,
 ):
     template_name = 'callisto_core/delivery/wizard_form.html'
     done_template_name = 'callisto_core/delivery/review.html'
-    storage_helper = EncryptedStorageHelper
-    steps_helper = ReportStepsHelper
+    storage_helper = view_helpers.EncryptedStorageHelper
+    steps_helper = view_helpers.ReportStepsHelper
 
     def dispatch(self, request, step=None, *args, **kwargs):
         self._dispatch_processing(step)
@@ -58,18 +24,21 @@ class EncryptedWizardView(
 
 
 class WizardActionView(
-    ReportActionView,
+    view_partials.ReportActionView,
     EncryptedWizardView,
 ):
+    # NOTE: this is technically a view_partial
+
     def dispatch(self, request, *args, **kwargs):
-        return super().dispatch(
-            request, step=StepsHelper.done_name, *args, **kwargs)
+        step = view_helpers.ReportStepsHelper
+        return super().dispatch(request, step=step, *args, **kwargs)
 
 
-class WizardPDFView(WizardActionView):
+class WizardPDFView(
+    WizardActionView,
+):
 
-    def _report_action(self):
-        # TODO: self.action.withdraw()
+    def _action_response(self):
         return self._report_pdf_response()
 
     def _report_pdf_response(self):
